Full Application with Standard GUI
The application does simple Crud functions and has a Login Username / Password hardcoded. 
The Login can support a database but I didn't need to add one for the open-source copy. 
Everything is in WPFAPP1 Namespace again because this is the Open Source Copy. 
The application has a search function as well.
User Page has not been created because I left it open to have another page create or link to the same page for the OPEN Source Verison. 
The Live version allows for users to have a separate view from the admins. 
The project took about 2 weeks to get to this state. I estimate the full Version to be complete and pushed live to my job in the next 2 weeks as well. 
If you use this. Make sure to change the DATABASE/SQL and LINQ functions to reflect your own database as then it will not work.